Minor reorganization of Scraper; update Scraper specs
Introduces a 'Client' class that includes HTTParty and then sets some
sane defaults for all of the lib's requests. I kept the Scraper module
in tact and just updated it to use this new Client class so that
`include RapGenius::Scraper` in Song and Annotation still worked.

Where possible I kept everything the same for backwards-compatibility
until @timrogers wants to change things.

Also updated are the specs for the module. Removed some specs that were
basically testing low-level HTTParty stuff.